# Add missing type information pieces