--- a/hilti/toolchain/include/ast/forward.h
+++ b/hilti/toolchain/include/ast/forward.h
@@ -93,7 +93,6 @@ class Assign;
 class BuiltInFunction;
 class Coerced;
 class Ctor;
-class Deferred;
 class Grouping;
 class Keyword;
 class ListComprehension;