A few pieces were still missing from the type information system so
that host applications ca get full, dynamic runtime access to all
values that HILTI can represent.

- **Report `ast-final` debugging output even if AST validation fails.**
- **Make pre-defined type information objects constant.**
- **Add a few missing type information pieces.**
- **Change FFI calling convention for `any`.**
- **Change runtime representation of tuples.**
- **Remove AST node for deferred expressions.**

This needs https://github.com/zeek/zeek/pull/4300 Zeek-side.