@rsmmr 2025-03-27
> This has a significant performance degradation on the unit microbenchmarks, just from running after rebasing on `main`

Ack, I can reproduce this. I'll see if I can get that down.