@rsmmr 2025-03-27
> Ack, I can reproduce this. I'll see if I can get that down.

@bbannier pointed out the initialization of `__result` variables, which are now more expensive, even though we actually don't need them. We cannot figure that out automatically right now but I added a mechanism to suppress the intialization manually. That now leaves it faster than before:

Branch:

```
benchmarkParser/Benchmark::UnitVectorSize/100              12535 ns        12536 ns        55573
benchmarkParser/Benchmark::UnitVectorSize/1000            107249 ns       107249 ns         6398
benchmarkParser/Benchmark::UnitVectorSize/10000          1199655 ns      1199652 ns          587
benchmarkParser/Benchmark::UnitVectorSize/100000        11470219 ns     11470081 ns           62
benchmarkParser/Benchmark::UnitVectorLookahead/100         17176 ns        17178 ns        40643
benchmarkParser/Benchmark::UnitVectorLookahead/1000       153268 ns       153259 ns         4524
benchmarkParser/Benchmark::UnitVectorLookahead/10000     1708999 ns      1708978 ns          410
benchmarkParser/Benchmark::UnitVectorLookahead/100000   16017453 ns     16017581 ns           43
benchmarkParser/Benchmark::Regex/100                        5211 ns         5213 ns       134141
benchmarkParser/Benchmark::Regex/1000                      39557 ns        39563 ns        17692
benchmarkParser/Benchmark::Regex/10000                    383195 ns       383202 ns         1820
benchmarkParser/Benchmark::Regex/100000                  3901545 ns      3900750 ns          184
```

Main:

```
benchmarkParser/Benchmark::UnitVectorSize/100              12899 ns        12894 ns        55137
benchmarkParser/Benchmark::UnitVectorSize/1000            113065 ns       113021 ns         6201
benchmarkParser/Benchmark::UnitVectorSize/10000          1170512 ns      1170276 ns          604
benchmarkParser/Benchmark::UnitVectorSize/100000        12768337 ns     12768389 ns           54
benchmarkParser/Benchmark::UnitVectorLookahead/100         18084 ns        18047 ns        38831
benchmarkParser/Benchmark::UnitVectorLookahead/1000       160701 ns       160634 ns         4317
benchmarkParser/Benchmark::UnitVectorLookahead/10000     1633655 ns      1633371 ns          428
benchmarkParser/Benchmark::UnitVectorLookahead/100000   17582041 ns     17582025 ns           40
benchmarkParser/Benchmark::Regex/100                        6699 ns         6706 ns       104140
benchmarkParser/Benchmark::Regex/1000                      51015 ns        51015 ns        13609
benchmarkParser/Benchmark::Regex/10000                    499000 ns       498962 ns         1412
benchmarkParser/Benchmark::Regex/100000                  4905601 ns      4905648 ns          142
```

Now the question is: does this pass review ....?