@rsmmr 2025-03-27
@bbannier I ended up leaving `get_all` in. Replacing that made the generated code quite a bit harder to read and didn't change that much in the end.