@rsmmr 2025-03-27
@evantypanski Could you try the updated branch with your SSL trace? What I'm seeing is very close, now about 2-3% overhead,