@evantypanski 2025-03-27
It's a lot closer on that trace, about 3% difference now

```
$ hyperfine "~/.local/zeek-current/bin/zeek -Cr out.trace" "~/.local/zeek-new-tup/bin/zeek -Cr out.trace" --warmup 3
Benchmark 1: ~/.local/zeek-current/bin/zeek -Cr out.trace
  Time (mean ± σ):      1.538 s ±  0.012 s    [User: 1.470 s, System: 0.113 s]
  Range (min … max):    1.521 s …  1.568 s    10 runs

Benchmark 2: ~/.local/zeek-new-tup/bin/zeek -Cr out.trace
  Time (mean ± σ):      1.589 s ±  0.011 s    [User: 1.519 s, System: 0.113 s]
  Range (min … max):    1.575 s …  1.608 s    10 runs

Summary
  ~/.local/zeek-current/bin/zeek -Cr out.trace ran
    1.03 ± 0.01 times faster than ~/.local/zeek-new-tup/bin/zeek -Cr out.trace
```